Extracted Shadowburn logic into module and added a statistic I wanted
