+++ src/Parser/DestructionWarlock/CombatLogParser.js
@@ -21,6 +21,7 @@ import DamageDone from './Modules/Features/DamageDone';
 
 import Backdraft from './Modules/Talents/Backdraft';
 import RoaringBlaze from './Modules/Talents/RoaringBlaze';
+import Shadowburn from './Modules/Talents/Shadowburn';
 import SoulHarvest from './Modules/Talents/SoulHarvest';
 import SoulHarvestTalent from './Modules/Talents/SoulHarvestTalent';
 
@@ -48,6 +49,7 @@ class CombatLogParser extends CoreCombatLogParser {
     //Talents
     backdraft: Backdraft,
     roaringBlaze: RoaringBlaze,
+    shadowburn: Shadowburn,
     soulHarvest: SoulHarvest,
     soulHarvestTalent: SoulHarvestTalent,
